Refactoring

See merge request spid/spid-saml-compliance-tests!4

+++ script/check-response-xsd-and-signature.sh
@@ -8,22 +8,22 @@ FROM=${2}
 
 SAMLResponse=`cat ${_DATA_DIR}/SAMLResponse.${CTX}.response.txt`
 if [ ${_DEBUG} -eq 1 ]; then
-    echo -e "\n[SAMLResponse]\n\n${SAMLResponse}"
+    echo -e "[SAMLResponse]\n${SAMLResponse}"
 fi
 
 RelayState=`cat ${_DATA_DIR}/RelayState.${CTX}.response.txt`
 if [ ${_DEBUG} -eq 1 ]; then
-    echo -e "\n[RelayState]\n\n${RelayState}"
+    echo -e "[RelayState]\n${RelayState}"
 fi
 
 Signature=`cat ${_DATA_DIR}/Signature.${CTX}.response.txt`
 if [ ${_DEBUG} -eq 1 ]; then
-    echo -e "\n[Signature]\n\n${Signature}"
+    echo -e "[Signature]\n${Signature}"
 fi
 
 SigAlg=`cat ${_DATA_DIR}/SigAlg.${CTX}.response.txt`
 if [ ${_DEBUG} -eq 1 ]; then
-    echo -e "\n[SigAlg]\n\n${SigAlg}\n"
+    echo -e "[SigAlg]\n${SigAlg}\n"
 fi
 
 if [ "X${Signature}" == "X" -a "X${SigAlg}" == "X" ]; then # HTTP-POST
@@ -31,17 +31,20 @@ if [ "X${Signature}" == "X" -a "X${SigAlg}" == "X" ]; then # HTTP-POST
     # decode SAMLResponse
     res=`mktemp`
     echo -n ${SAMLResponse} | base64 -d > ${res}
-    
+
     if [ ${_DEBUG} -eq 1 ]; then
-        echo -e "\n[SAMLResponse]\n\n`xmllint --format ${res}`"
+        echo -e "[SAMLResponse]\n`xmllint --format ${res}`"
     fi
 
     # verify against XSD
+    echo -n "Validating XSD... "
     xmllint --noout --schema ./xsd/saml-schema-protocol-2.0.xsd ${res}
     if [ $? -ne 0 ]; then
+        echo "FAIL"
         rm ${res}
         exit 1
     fi
+    echo "OK"
 
     # verify response XML signature (outer)
     if [ "${CTX}" == "authn" ]; then
@@ -52,6 +55,8 @@ if [ "X${Signature}" == "X" -a "X${SigAlg}" == "X" ]; then # HTTP-POST
         rm ${res}
         exit 1
     fi
+
+    echo -n "Validating response signature... "
     xmlsec1 \
         --verify \
         --insecure \
@@ -59,18 +64,25 @@ if [ "X${Signature}" == "X" -a "X${SigAlg}" == "X" ]; then # HTTP-POST
         --node-name ${elem} \
         ${res}
     if [ $? -ne 0 ]; then
+        echo "FAIL"
         rm ${res}
         exit 1
     fi
+    echo "OK"
 
-    # verify assertion XML signature (inner)
-    xmlsec1 --verify --insecure \
-        --id-attr:ID urn:oasis:names:tc:SAML:2.0:assertion:Assertion \
-        --node-name urn:oasis:names:tc:SAML:2.0:assertion:Assertion \
-        ${res}
-    if [ $? -ne 0 ]; then
-        rm ${res}
-        exit 1
+    if [ "${CTX}" == "authn" ]; then
+        # verify assertion XML signature (inner)
+        echo -n "Validating assertion signature... "
+        xmlsec1 --verify --insecure \
+            --id-attr:ID urn:oasis:names:tc:SAML:2.0:assertion:Assertion \
+            --node-name urn:oasis:names:tc:SAML:2.0:assertion:Assertion \
+            ${res}
+        if [ $? -ne 0 ]; then
+            echo "FAIL"
+            rm ${res}
+            exit 1
+        fi
+        echo "OK"
     fi
 
     rm ${res}
@@ -83,17 +95,20 @@ else # HTTP-Redirect
         | python -c "import sys, zlib;\
                      sys.stdout.write(zlib.decompress(sys.stdin.buffer.read(), -15).decode('utf-8'))" \
         > ${res}
-    
+
     if [ ${_DEBUG} -eq 1 ]; then
-        echo -e "\n[SAMLResponse]\n\n`xmllint --format ${res}`"
+        echo -e "[SAMLResponse]\n`xmllint --format ${res}`"
     fi
-    
+
     # verify against XSD
+    echo -n "Validating XSD... "
     xmllint --noout --schema ./xsd/saml-schema-protocol-2.0.xsd ${res}
     if [ $? -ne 0 ]; then
+        echo "FAIL"
         rm ${res}
         exit 1
     fi
+    echo "OK"
 
     # decode XML signature
     signature=`mktemp`
@@ -103,11 +118,11 @@ else # HTTP-Redirect
     response=`echo -n ${SAMLResponse} \
         | python -c "import sys, urllib.parse as p;\
                      sys.stdout.write(p.quote_plus(sys.stdin.read()))"`
-    
+
     relay_state=`echo -n ${RelayState} \
         | python -c "import sys, urllib.parse as p;\
                      sys.stdout.write(p.quote_plus(sys.stdin.read()))"`
-    
+
     sig_alg=`echo -n ${SigAlg} \
         | python -c "import sys, urllib.parse as p;\
                      sys.stdout.write(p.quote_plus(sys.stdin.read()))"`
@@ -130,10 +145,26 @@ else # HTTP-Redirect
         rm ${pubkey}
     done
 
+    echo -n "Validating response signature... "
     if [ ${sign_ok} -ne 1 ]; then
+        echo "FAIL"
         rm ${payload} ${signature} ${res}
         exit 1
     fi
-    
+    echo "OK"
+
+    # verify assertion XML signature (inner)
+    echo -n "Validating assertion signature... "
+    xmlsec1 --verify --insecure \
+        --id-attr:ID urn:oasis:names:tc:SAML:2.0:assertion:Assertion \
+        --node-name urn:oasis:names:tc:SAML:2.0:assertion:Assertion \
+        ${res}
+    if [ $? -ne 0 ]; then
+        echo "FAIL"
+        rm ${res}
+        exit 1
+    fi
+    echo "OK"
+
     rm ${payload} ${signature} ${res}
 fi
